+++ doc/whats-new.rst
@@ -39,6 +39,11 @@ Enhancements
 Bug fixes
 ~~~~~~~~~
 
+- Fixed ``DataArray.to_iris()`` failure while creating ``DimCoord`` by 
+  falling back to creating ``AuxCoord``. Fixed dependency on ``var_name``
+  attribute being set.
+  (:issue:`2201`)
+  By `Thomas Voigt <https://github.com/tv3141>`_.
 
 .. _whats-new.0.10.8: